requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.3.4
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
apio==0.8.4
asgiref==3.8.1
astrapy==1.1.0
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
bson==0.5.10
build==1.2.1
cachetools==5.4.0
cassandra-driver==3.29.2
cassio==0.1.8
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
Chroma==0.2.0
chroma-hnswlib==0.7.6
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
daal==2024.1.0
daal4py==2024.1.0
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
executing==2.0.1
fastapi==0.115.0
fastjsonschema==2.20.0
ffmpy==0.3.2
filelock==3.13.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
flatbuffers==24.3.25
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
gradio==4.22.0
gradio_client==0.13.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.29.3
ipython==8.22.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.5.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.3.1
langchain-astradb==0.2.0
langchain-core==0.3.12
langchain-text-splitters==0.3.0
langchainhub==0.1.21
langsmith==0.1.136
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.0
marshmallow==3.22.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.9.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
neo4j==5.25.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==0.28.0
openapi==2.0.0
opencv-python==4.9.0.80
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.3
pdf2image==1.17.0
pika==1.3.2
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.2.0
playwright==1.47.0
posthog==3.7.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.0.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
pymongo==4.10.1
PyMySQL==1.1.1
pyparsing==3.1.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.2
redis==5.1.1
referencing==0.35.1
regex==2024.9.11
reportlab==4.2.5
requests==2.32.2
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.1
rpds-py==0.20.0
rsa==4.9
ruff==0.3.4
safetensors==0.4.2
scikit-learn==1.4.1.post1
scikit-learn-intelex==2024.1.0
scipy==1.12.0
SCons==4.2.0
seaborn==0.13.2
semantic-version==2.9.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlearn==0.0.1
stack-data==0.6.3
starlette==0.38.6
streamlit==1.39.0
sympy==1.13.3
tavily-python==0.5.0
tbb==2021.11.0
tenacity==8.5.0
text-unidecode==1.3
threadpoolctl==3.3.0
tiktoken==0.8.0
timm==0.9.16
tinycss2==1.3.0
tldextract==5.1.2
tokenizers==0.20.0
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.1
typer==0.12.5
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uuid6==2024.1.12
uvicorn==0.31.0
vboxapi==1.0
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.3
wheel==0.42.0
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.13.1
zipp==3.20.2
langchain-community
langchain-openai
langchain-chroma
chromadb>=0.4.0